+++ migration/env.py
@@ -7,7 +7,7 @@ from alembic import context
 from src.di.container import DIContainer
 from src.infrastructure.repository.sqlalchemy.config import SQLAlchemyConfigIF
 from src.infrastructure.repository.sqlalchemy.model.all import MODELS
-from src.infrastructure.repository.sqlalchemy.model.baes import Base
+from src.infrastructure.repository.sqlalchemy.model.base import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
