+++ src/infrastructure/repository/sqlalchemy/model/all.py
@@ -1,6 +1,6 @@
 from typing import Final
 
-from src.infrastructure.repository.sqlalchemy.model.baes import Base
+from src.infrastructure.repository.sqlalchemy.model.base import Base
 from src.infrastructure.repository.sqlalchemy.model.messenger import Messenger
 from src.infrastructure.repository.sqlalchemy.model.nickname import NicknameChangelog
 from src.infrastructure.repository.sqlalchemy.model.user import User
